Moving fake implementations to a common structure
