+++ TeamCode/src/test/java/com/hfrobots/tnt/corelib/control/KonamiCodeTest.java
@@ -25,6 +25,7 @@ import com.hfrobots.tnt.corelib.state.DelayState;
 import com.hfrobots.tnt.corelib.state.State;
 import com.hfrobots.tnt.fakes.FakeNinjaGamePad;
 import com.hfrobots.tnt.fakes.FakeTelemetry;
+import com.hfrobots.tnt.fakes.control.FakeOnOffButton;
 
 import org.junit.Assert;
 import org.junit.Before;
