+++ TeamCode/src/test/java/com/hfrobots/tnt/season1819/ParticleScoringMechanismTest.java
@@ -1,9 +1,9 @@
 package com.hfrobots.tnt.season1819;
 
 import com.hfrobots.tnt.corelib.control.DebouncedButton;
-import com.hfrobots.tnt.corelib.control.FakeOnOffButton;
-import com.hfrobots.tnt.corelib.drive.FakeExtendedDcMotor;
-import com.hfrobots.tnt.corelib.drive.FakeServo;
+import com.hfrobots.tnt.fakes.control.FakeOnOffButton;
+import com.hfrobots.tnt.fakes.drive.FakeExtendedDcMotor;
+import com.hfrobots.tnt.fakes.drive.FakeServo;
 import com.hfrobots.tnt.fakes.FakeTelemetry;
 
 import junit.framework.Assert;
